Add {non}synonymous, update excel, change rbind
Filter mutations in synonymous and nonsynonymous in addition to the already
existing high_moderate and low_modifier postfixes. All mutations are classified
as nonsynonymous if they have high/moderate impact according to snpeff. The low
impact category ends up in the synonymous category if it is a
synonymous_variant. Note that there are still some nonsynonymous start/stop
mutations in the low impact category that are not included. Also, notice that
by definition {non}synonymous mutations are only in the exon region so all
introns are excluded. The rbind script has been changed to output just the
header for alltables in case there are no mutations.