Add recurrent mutations sufam report
Run sufam on the set of all NS mutations over all samples and generate an
ipython notebook report that shows the maf of each recurrent mutation. Plots
are generated for a mutation at the gene level and an identical mutation in
multiple samples. Parameters for the plots can be set by changing config.inc.
At the moment one can choose in how many samples the mutation must occur and
which samples and in what order they should be plotted.