Use hotspot v1/v2 union
Update variant_calls test files and point to new hotspot directory. The
hotspots are a union of v1 and MSKCC internal v2 (soon to be published). The v1
is on TCGA exome cases, whereas v2 is on a larger IMPACT gene panel of IMPACT
cases. That is one reason not all hotspots of v1 are in v2, Other reasons could be :

1. different tumor type proportions in the IMPACT cases
2. Biological false positives.T hese v1 hotspots were called because of a smaller sample size and with more samples in v2 hotspots it is no longer significant
3. Technical false positive. These can be germline and/or sequencing artifacts that are present in the TCGA calls but not in IMPACT cases